requirements.txt
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.8.0
arize-phoenix==8.26.1
arize-phoenix-client==1.3.0
arize-phoenix-evals==0.20.4
arize-phoenix-otel==0.9.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
Authlib==1.5.2
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
cryptography==44.0.2
dataclasses-json==0.6.7
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==7.4.3
e2b==1.3.4
e2b-code-interpreter==1.2.0
email_validator==2.2.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
geographiclib==2.0
geopandas==1.0.1
googleapis-common-protos==1.69.0
gradio==5.16.1
gradio_client==1.7.0
graphql-core==3.2.6
greenlet==3.1.1
grpc-interceptor==0.15.4
grpcio==1.70.0
h11==0.14.0
h2==4.2.0
helium==5.1.1
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1738421264/work
ipywidgets==8.1.5
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_widgets==3.0.13
kaleido==0.2.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.30
litellm==1.66.0
lxml==5.3.1
Mako==1.3.10
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mcp==1.6.0
mcpadapt==0.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
openai==1.74.0
openinference-instrumentation==0.1.23
openinference-instrumentation-smolagents==0.1.6
openinference-semantic-conventions==0.1.14
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
plotly==6.0.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
propcache==0.3.1
protobuf==5.29.4
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyogrio==0.10.0
pyproj==3.7.1
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1734687138743/work
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
semantic-version==2.10.0
setuptools==75.8.0
shapely==2.1.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smolagents==1.9.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlean.py==3.47.0
sse-starlette==2.2.1
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.45.3
strawberry-graphql==0.265.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado @ file:///croot/tornado_1733960490606/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.19.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
